+++ analysis/EventAnalysis.hh
@@ -100,6 +100,13 @@ private:
   /// Process each sampler analysis object.
   void ProcessSamplers(bool firstTime = false);
 
+  /// The data is different for different sampler types and therefore we must
+  /// specialise the PerEnetryHistogramSet. This delagator function constructs
+  /// the right one.
+  PerEntryHistogramSet* _ConstructPerEntryHistogramSet(const HistogramDefSet* definitionIn,
+                                                       Event*                 eventIn,
+                                                       TChain*                chainIn) const;
+
   bool printOut;          ///< Whether to print out at all per-event.
   int  printModulo;       ///< Cache of print modulo fraction
   bool processSamplers;   ///< Whether to process samplers.