+++ analysis/PerEntryHistogramSet.cc
@@ -95,16 +95,6 @@ PerEntryHistogramSet::~PerEntryHistogramSet()
     {delete kv;}
 }
 
-void PerEntryHistogramSet::CheckSampler()
-{
-  if (!sampler)
-    {
-      sampler = event->GetSampler(branchName + "."); // cache sampler
-      if (!sampler)
-        {throw RBDSException("Cannot find sampler \"" + branchName + "\" or Model tree was not stored");}
-    }
-}
-
 void PerEntryHistogramSet::AccumulateCurrentEntry(long int entryNumber)
 {
   CheckSampler();
@@ -113,7 +103,8 @@ void PerEntryHistogramSet::AccumulateCurrentEntry(long int entryNumber)
     {
       // for this event, form a set of pdgIDs and
       // then ensure we have all prepare histograms
-      std::set<long long int> pdgIDSet(sampler->partID.begin(), sampler->partID.end());
+      std::set<long long int> pdgIDSet;
+      GetPDGIDSetFromSampler(pdgIDSet);
 
       // use vector because set cannot be appended to for set_difference
       std::vector<long long int> missing;